pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "transcript-cleaner"
version = "0.1.0"
description = "Fix messy interview transcripts — speaker labels, fillers, fragmentation, timestamps — with rule-based heuristics and optional LLM assistance."
readme = "README.md"
license = { file = "LICENSE" }
authors = [{ name = "Kostas Kazakos", email = "kgkazakos@gmail.com" }]
keywords = ["ux-research", "transcripts", "nlp", "llm", "qualitative-research", "cpr"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Text Processing :: Linguistic",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
requires-python = ">=3.10"
dependencies = []

[project.optional-dependencies]
cli = ["click>=8.1"]
openai = ["openai>=1.0"]
anthropic = ["anthropic>=0.25"]
gemini = ["google-generativeai>=0.5"]
docx = ["python-docx>=1.0"]
all = [
    "click>=8.1",
    "openai>=1.0",
    "anthropic>=0.25",
    "google-generativeai>=0.5",
    "python-docx>=1.0",
]
dev = [
    "pytest>=7.0",
    "pytest-cov",
    "ruff",
    "mypy",
]

[project.scripts]
transcript-cleaner = "transcript_cleaner.cli:main"

[project.urls]
Homepage = "https://github.com/kgkazakos/transcript-cleaner"
Repository = "https://github.com/kgkazakos/transcript-cleaner"
"Bug Tracker" = "https://github.com/kgkazakos/transcript-cleaner/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["transcript_cleaner*"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=transcript_cleaner --cov-report=term-missing"
